+++ hydra/symmetries/group_action/group_action.cpp
@@ -12,12 +12,12 @@ GroupAction::GroupAction(PermutationGroup const &permutation_group)
       permutation_group_(permutation_group), indices_(n_symmetries_, 0),
       fermi_work_(2 * permutation_group.n_sites(), 0) {}
 
-template <class bit_t> bit_t GroupAction::apply(int sym, bit_t state) const {
+template <class bit_t> bit_t GroupAction::apply(int64_t sym, bit_t state) const {
   return permutation_group_[sym].apply(state);
 }
-template uint16_t GroupAction::apply<uint16_t>(int, uint16_t) const;
-template uint32_t GroupAction::apply<uint32_t>(int, uint32_t) const;
-template uint64_t GroupAction::apply<uint64_t>(int, uint64_t) const;
+template uint16_t GroupAction::apply<uint16_t>(int64_t, uint16_t) const;
+template uint32_t GroupAction::apply<uint32_t>(int64_t, uint32_t) const;
+template uint64_t GroupAction::apply<uint64_t>(int64_t, uint64_t) const;
 
 template <class bit_t> bit_t GroupAction::representative(bit_t state) const {
   bit_t rep = std::numeric_limits<bit_t>::max();
@@ -35,11 +35,11 @@ template uint32_t GroupAction::representative<uint32_t>(uint32_t) const;
 template uint64_t GroupAction::representative<uint64_t>(uint64_t) const;
 
 template <class bit_t>
-std::pair<bit_t, int> GroupAction::representative_sym(bit_t state) const {
+std::pair<bit_t, int64_t> GroupAction::representative_sym(bit_t state) const {
   bit_t rep = std::numeric_limits<bit_t>::max();
-  int idx = 0;
+  int64_t idx = 0;
 
-  int sym=0;
+  int64_t sym=0;
   for (auto const &p : permutation_group_) {
     bit_t trans = p.apply(state);
     if (trans < rep) {
@@ -50,20 +50,20 @@ std::pair<bit_t, int> GroupAction::representative_sym(bit_t state) const {
   }
   return {rep, idx};
 }
-template std::pair<uint16_t, int>
+template std::pair<uint16_t, int64_t>
     GroupAction::representative_sym<uint16_t>(uint16_t) const;
-template std::pair<uint32_t, int>
+template std::pair<uint32_t, int64_t>
     GroupAction::representative_sym<uint32_t>(uint32_t) const;
-template std::pair<uint64_t, int>
+template std::pair<uint64_t, int64_t>
     GroupAction::representative_sym<uint64_t>(uint64_t) const;
 
 template <class bit_t>
-std::pair<bit_t, gsl::span<int const>>
+std::pair<bit_t, gsl::span<int64_t const>>
 GroupAction::representative_syms(bit_t state) const {
   bit_t rep = std::numeric_limits<bit_t>::max();
-  gsl::span<int const>::size_type n_indices = 0;
+  gsl::span<int64_t const>::size_type n_indices = 0;
 
-  int sym=0;
+  int64_t sym=0;
   for (auto const &p : permutation_group_) {
     bit_t trans = p.apply(state);
     if (trans < rep) {
@@ -78,31 +78,31 @@ GroupAction::representative_syms(bit_t state) const {
   return {rep, {indices_.data(), n_indices}};
 }
 
-template std::pair<uint16_t, gsl::span<int const>>
+template std::pair<uint16_t, gsl::span<int64_t const>>
     GroupAction::representative_syms<uint16_t>(uint16_t) const;
-template std::pair<uint32_t, gsl::span<int const>>
+template std::pair<uint32_t, gsl::span<int64_t const>>
     GroupAction::representative_syms<uint32_t>(uint32_t) const;
-template std::pair<uint64_t, gsl::span<int const>>
+template std::pair<uint64_t, gsl::span<int64_t const>>
     GroupAction::representative_syms<uint64_t>(uint64_t) const;
 
 template <class bit_t>
-double GroupAction::fermi_sign(int sym, bit_t state) const {
+double GroupAction::fermi_sign(int64_t sym, bit_t state) const {
   return symmetries::fermi_sign_of_permutation(state, permutation_group_[sym],
                                                fermi_work_);
 }
-template double GroupAction::fermi_sign<uint16_t>(int, uint16_t) const;
-template double GroupAction::fermi_sign<uint32_t>(int, uint32_t) const;
-template double GroupAction::fermi_sign<uint64_t>(int, uint64_t) const;
+template double GroupAction::fermi_sign<uint16_t>(int64_t, uint16_t) const;
+template double GroupAction::fermi_sign<uint32_t>(int64_t, uint32_t) const;
+template double GroupAction::fermi_sign<uint64_t>(int64_t, uint64_t) const;
 
 template <class bit_t>
-std::vector<int> GroupAction::stabilizer_symmetries(bit_t bits) const {
+std::vector<int64_t> GroupAction::stabilizer_symmetries(bit_t bits) const {
   return symmetries::stabilizer_symmetries(bits, *this);
 }
-template std::vector<int>
+template std::vector<int64_t>
     GroupAction::stabilizer_symmetries<uint16_t>(uint16_t) const;
-template std::vector<int>
+template std::vector<int64_t>
     GroupAction::stabilizer_symmetries<uint32_t>(uint32_t) const;
-template std::vector<int>
+template std::vector<int64_t>
     GroupAction::stabilizer_symmetries<uint64_t>(uint64_t) const;
 
 bool GroupAction::operator==(GroupAction const &rhs) const {
