+++ test/symmetries/group_action/test_group_action.cpp
@@ -8,16 +8,16 @@
 
 using namespace hydra;
 
-template <class bit_t> void test_permutation_group_action(int n_sites) {
+template <class bit_t> void test_permutation_group_action(int64_t n_sites) {
   using combinatorics::Subsets;
 
   // test cyclic group
   std::vector<Permutation> permutation_array;
-  for (int sym = 0; sym < n_sites; ++sym) {
+  for (int64_t sym = 0; sym < n_sites; ++sym) {
 
-    std::vector<int> pv;
-    for (int site = 0; site < n_sites; ++site) {
-      int newsite = (site + sym) % n_sites;
+    std::vector<int64_t> pv;
+    for (int64_t site = 0; site < n_sites; ++site) {
+      int64_t newsite = (site + sym) % n_sites;
       pv.push_back(newsite);
     }
     permutation_array.push_back(Permutation(pv));
@@ -30,7 +30,7 @@ template <class bit_t> void test_permutation_group_action(int n_sites) {
   // for (auto state : Subsets<bit_t>(n_sites)) {
   //   std::cout << bits_to_string(state, n_sites) << "\n";
 
-  //   for (int sym = 0; sym < n_sites; ++sym) {
+  //   for (int64_t sym = 0; sym < n_sites; ++sym) {
   //     auto tstate = sym_op.apply(sym, state);
   //     std::cout << sym << " " << bits_to_string(tstate, n_sites) << "\n";
   //   }
@@ -39,8 +39,8 @@ template <class bit_t> void test_permutation_group_action(int n_sites) {
 
   // Check whether cyclic group property holds for every state
   for (auto state : Subsets<bit_t>(n_sites)) {
-    for (int sym1 = 0; sym1 < n_sites; ++sym1)
-      for (int sym2 = 0; sym2 < n_sites; ++sym2) {
+    for (int64_t sym1 = 0; sym1 < n_sites; ++sym1)
+      for (int64_t sym2 = 0; sym2 < n_sites; ++sym2) {
         auto tstate1 = sym_op.apply(sym1, state);
         auto tstate2 = sym_op.apply(sym2, tstate1);
         auto tstate12 = sym_op.apply((sym1 + sym2) % n_sites, state);
@@ -57,7 +57,7 @@ template <class bit_t> void test_permutation_group_action(int n_sites) {
     // std::cout << "r " << bits_to_string(rep, n_sites) << "\n\n";
 
     std::vector<bit_t> orbit;
-    for (int sym = 0; sym < n_sites; ++sym) {
+    for (int64_t sym = 0; sym < n_sites; ++sym) {
       orbit.push_back(sym_op.apply(sym, state));
     }
     auto min = std::min_element(orbit.begin(), orbit.end());
@@ -70,7 +70,7 @@ template <class bit_t> void test_permutation_group_action(int n_sites) {
     auto [rep3, syms] = sym_op.representative_syms(state);
     REQUIRE(rep3 == rep);
 
-    for (int sym : syms) {
+    for (int64_t sym : syms) {
       auto tstate = sym_op.apply(sym, state);
       REQUIRE(rep == tstate);
     }
@@ -79,7 +79,7 @@ template <class bit_t> void test_permutation_group_action(int n_sites) {
 
 TEST_CASE("GroupAction", "[symmetries]") {
   hydra::Log("Test GroupAction");
-  for (int n_sites = 1; n_sites < 6; ++n_sites) {
+  for (int64_t n_sites = 1; n_sites < 6; ++n_sites) {
     test_permutation_group_action<uint16_t>(n_sites);
     test_permutation_group_action<uint32_t>(n_sites);
     test_permutation_group_action<uint64_t>(n_sites);
